Add logrus usage to agent
To monitor operations between agent and the service it require to keep
using the same request-id on each interaction.
Added a reqid package to have a standard request id format and a way to
get a new request id when needed.
Sequential request id will better to understand the flow of the
interaction
Instead of passing the logger to each function device logger is updated
on each interaction with the request id field.

In the future if there will be async operations on the agent it will
require to pass the logger with the relevant request id to the async
part in order to trace the operations

I decided not to use the context to keep the request id because the way
`WithValue` function is implemented, it will add additional context each
time with value is called in the long run it will create a long chane
of value context objects

To set the context from the agent i added a request editor function
(WithRequestEditorFn) that update the header and setting the request id,
this way each interaction is tracked with the same request id between
the client and the server