Create task infra based on channels
The task infra design:
- A goroutine for each task
- The task is notified when a resource (or part of it) is updated via
  channel

This is a temporary solution until we decide on a task tool. This task
infra does not handle retries and tasks are lost when the server is
terminated.

Also translated the "device updater" periodic task to use this infra.
The new task is called "fleet rollout", and is triggered whenever a
fleet's template is updated. This doesn't handle devices moving from one
fleet to another - that will be handled in a separately once the service
tracks what fleet each device belongs to.