MGMT-17037: Update device owner based on fleet label selector
We have 4 cases:
1. Fleet with no overlapping selectors, create/update/delete:
   Iterate devices that match the fleet's selector and set
   owners/conditions as necessary.
2. Fleet with overlapping selectors, create/update/delete:
   Iterate all fleets and devices in the org and set owners/conditions
   as necessary.
3. Device with a single owner, create/update (no work needed for delete):
   Iterate fleets and set the device's owner/conditions as necessary.
4. Device with multiple owners, create/update/delete:
   Iterate all fleets and devices in the org and set owners/conditions
   as necessary.

This commit introduces Annotations in ObjectMeta. These are key/values
set and used by the service. In this case, we have an annotation on
devices that have multiple owners to differentiate between cases 3 and 4
above.

Fleets with overlapping selectors have an "OverlappingSelectors"
Conditition set by the service.

In addition, we have the cases where the user deleted all fleets or
devices in an org. In this case we unset any set conditions/annotations.