MGMT-16972: Deploy via kind+helm
Enable deployment from kind+helm, replacing the podman deployment
method.

While we want to ensure that no hard K8s or OpenShift dependencies
are introduced (and that was the main reason for podman), maintaing
two sets of deployment files was error prone, and does't allow for
testing the optional Secrets/ConfigFile dependencies eventually.

Signed-off-by: Miguel Angel Ajo Pelayo <majopela@redhat.com>