MGMT-17158: Fixed a crash in fleet store
Fixed `panic: runtime error: invalid memory address or nil pointer dereference` that was causing fleets and RS to get out of sync.

Also, if sync needs to happen, we now set an "out of sync" condition to prevent false positive sync status when crashes or bugs like this happen in the future.

This "out of sync" status will ensure that the process continues in the next task iteration.

Signed-off-by: Sagi Dayan <sagidayan@gmail.com>