NO-ISSUE: Gather client cert CN and create separate certs
From the http server, every time a connection is served
make sure that we gather the TLS client certificate details
and put the CommonName in the context, so it's accessible
later down when serving the API request.

Additionally when creating client certificates, define
different common names for a client-admin key/cert, and
for general agent bootstrapping certificates in a way
that we can differentiate the origin of requests.