NO-ISSUE: Optimize flightctl-server-container build
The go-cache in bin/go-cache creates issues with the linter and other
processes. And Copying the whole directory is slow when we have
artifacts in bin/

We could use dockerignore and be less explicit but then we cannot
copy some artifacts that we need from bin/ for other container builds.