OCTOET-412: Create several image variants for E2E and a local CA
Creates several image variants with dummy E2E services, some
crashing and some not, and push them to the locally hosted
registry. This can then be used by the agent to upgrade
to different versions.

Some services like bootc didn't have support for insecure tls
so we provide a local CA that the agents will trust and we
can put on services like the registry or git servers in the future.

 with '#' will be ignored, and an empty message aborts the commit.