MGMT-18036: Fix fetching git hashes
The code supported cloning branches and tags, but there is no concept in
git for cloning a hash. Fixed the code to perform a "checkout" after
clone in case a hash is requested.

Note that until now there was an optimization of passing "depth=1" which
means pulling with no history, but this means we can't check out hashes.
We will optimize later when implementing a cache.