OCTOET-476: Create agent api/client definitions
The openapi.yaml for the agent references the main one for schemas.

Subsequent commits will remove some of the methods and paths from
the general server, and move those into the agent only api.

OCTOET-476: update to oapi-gen v2.3.0

We reference the existing methods in the main openapi.yaml,
but this is broken in oapi-gen v1.xx, v2.3.0 improves
generation a bit but some details are still broken
(we apply workaround).