MGMT-18289: Support k8s secrets as a configuration source
K8S secrets are defined as config item in either fleet or device.  Such
config item contains the following fields:
- name: the secret name
- namespace: the secret namespace
- mountPoint: the location were the secret(s) should be stored.
Every secret has a "data" dictionary.  The data itself is stored as
key/value pairs.
The aim is to place the values on the filesystem of a target device.
So the mountPath is the parent directory, and each of the keys are the
file names in the parent directory.
Every K8S secret contains the values encoded with base64.  When placing
in the target device this encoding is removed.
To place the secrets on a target device, ignition is used.  Since
kubernetes config item is not in ignition format,
"InlineConfigProviderSpec" is used for that purpose.