MGMT-18494: Add hooks Manager
Add hooks manager that contains hooks that are invoked upon events.
The supported events are: before-create, after-create, before-update,
after-update, before-remove and after-remove.
The hooks-manager is a data structure that manges the hooks and "knows"
how to act upon the events.
This commit contains default hools for podman-compose.
The watched directory for podman-compose id /var/run/flightctl/compose.
The docker-compose invokes podman containers.
The watched directory for quadlets is /etc/containers/systemd.
This commit contains also support for removal.  The list of the expected
files are taken from the current configuration.