EDM-307: PATCH operations may encounter resourceVersion conflicts
The semantics of PATCH are that it overwrites specific fields without
regard for the previous contents. The current implementation fetches the
current resource from the DB, applies the patch, and calls CreateOrUpdate.
However, the fetched resource includes resourceVersion, and if someone
else changes the resource during this time, the resourceVersion will no
longer match and we'll return an error to the user.

The solution is to nil out the resourceVersion before calling
CreateOrUpdate so that it applies the patch regardless of the current
contents.