EDM-305: Fleet rollout doesn't retry on resourceVersion conflict
In fleet rollout, we fetch all devices in a fleet, and iterate through
them one-by-one. However, a device may change between fetching and
updating. This would currently cause a conflict error that isn't
retried.

In case of conflict, fetch the device resource again. If the owner
changed, skip the rollout. Otherwise, retry the update.