EDM-344: Fleet valid status is not set to 'false' if a repository present in its template config is deleted
Currently there is a cascade delete foreign key constraint on table
"fleet_repos" which is used for the many to many relationship between
table repositories and fleets.  This causes the association record to be
deleted when repository is deleted (which is not desirable).
This change removes the constraint during migration.  It means that the
association will only be deleted when the fleet is deleted.
Same change was done for table "device_repos".