EDM-398: A device becomes degraded after a compose file is removed
If there is an attempt to remove a compose file twice, running the hook
the second time may fail since the compose file does not exist.
This commit checks in case there is a failure to run the podman-compose
down, if the compose file does not exist and there is no container that
was launched by the compose file.
In additon, it clears degraded status for hooks, if a file a hook for a
failed file has been recovered.