EDM-513: Enable E2E on a target OpenShift deployment
This commit enables running the E2E tests on an existing OpenShift
cluster by using the ìn-cluster-e2e-test make target.

It expects two environment variables:

FLIGHTCTL_NS  which defaults to flightctl
KUBEADMIN_PASS  which is used to create tokens and login in automation

It also expects a kubeconfig to contact the target cluster.

In addition, a Vagrantfile for running the tests and development
environment has been included, this is helpful when the host VM
cannot run latest podman or other development tools necessary to
perform testing. Please note that this vagrant file only supports
x86_64 for now.