EDM-520: Implement rollout policy batch selection
The batch selection is implemented as periodic reconciler which is
invoked periodically by the periodic server.

The reconciler manages the whole flow of the batch sequence, including
iterating over batches, selecting devices for a batch, batch approval flow,
and sending a batch for rollout.

To avoid handling specific devices by this reconciler, its logic works
with aggregated counts.

EDM-524: Implement disruption allowance

The disruption allowance is also implemented as periodic reconciler
which is invoked by the periodic server.

The reconciler queries fleets that have devices that their spec needs to
be rendered.  It sends these devices to rendering according to the
restriction defined in the disruption allowance part in the spec.

EDM-523: Merge the rollout batch selection and disruption allowance to existing flow

The device selection and the disruption allowance are implemnted by
stopping flows executed as part of the existing tasks.  These flows are
fleet rollout (for device selection) and device render (for disruption
allowance).
When applicable, these operations are done by the reconcilers when the
required conditions are met.