EDM-1054: Remove ownership change check in store
A check for ownership change was added when creating the generic update
methods. This check is not necessary because the owner property is set
to nil at the service level of all resources to prevent users from
modifying it. Because the service layer sets the owner to nil, the check
fails for any resource with an owner because the code compares the nil
owner to the existing owner and thinks that the owner is being modified.
The solution is to remove the redundant check, which is the way it was
before the generic methods were introduced.