EDM-945: Return the most relevant openapi error
The openapi validations can return multiple errors. Not all of them are
relevant and which can be more noisy than helpful. This commit adds a
custom handler with a heuristic to highlight the error found at the
longest path, and return a short version of the others.

For example, the output in the bug report contains 4 errors, each with a
full schema and passed value:
Error at "/spec/config/1/gitRef": property "gitRef" is missing
Error at "/spec/config/1/secretRef": property "secretRef" is missing
Error at "/spec/config/1/inline/0/content": property "content" is missing
Error at "/spec/config/1/httpRef": property "httpRef" is missing

This commit changes the behavior to return the entire 3rd error, which
has the longest path, and mention the other 3 as above, without the
schema and value.