EDM-1062: Handle rollout of a new device or a device that just joined a fleet
In fleet-rollout task, stop rolling a device when there is a request to
roll it out if it is part of a fleet and there is a rollout in progress.
For batch sequence it checks if there is an active batch number and it
is less than the number of batches in the sequence.