EDM-1141: In some cases the rollout fails but it is not reported with an error message and does not timeout
When using disruption budget with minAvailable >= the total devices in a
group, then all devices in that group will be starved from rollout.  It
means that no rollout will take place for any device in this group.

This change verifies that the enforced minAvailable value will also be
less than the total devices in a group, so at least 1 device could be
rolled out for a device at a given time.