NO-ISSUE: Console sessions are not terminated in case a close request arrives from server
If a close request arrives from server, the bash process may be left
alive in the agent.
This change causes the bash process to terminate and clean up the
session.
Also, when a websocket is closed and the remote stream running on the
agent has not been closed, make sure that the reader from the stream
terminates by returning from the gRPC call when the websocket is closed.