EDM-1627: A clear error message is missing when we try to start a console session with a not existing device name
With the current kubernetes API, a clear error message is not extracted
from the websocket upgrade error.
This is fixed from kubernetes api 0.32.0 or later.
An error is now created by trying to perform "get" operation on the
device.  If an error occurs it is displayed.
Since the generated error is not necessarily the real error the occured,
the solution for this bug should be removed when moving to kubernetes
API 0.32.0 or later.