EDM-1793: Ensure event names conform to k8s standards
Resources in Flight Control API have a "metadata.name" which adhere to
the restrictions of k8s API. We were using the requestID as part of the
event name so that events can be easily correlated with requests, but
requestIDs sometimes had illegal characters like slashes and upper-case
letters.

The solution is to use a UUID for the event's name, and store the
requestID in an annotation.