EDM-1766: RPM/SELinux build tweaks/docs
The most important thing here is adding the conditional definition of the cgroup_type
to the SELinux policy, as it is not defined in either Centos 9 or 10 but is present
in the Fedora distros used to build RPMs. This was preventing the policy
module from installing on Centos. An e2e test to confirm that the policy
module was installed is also added.

Also adding several rules to allow the e2e tests to pass. They all seem
like reasonable things that a real user might want to do.

The rest is just docs/script changes to help avoid some of the pitfalls I ran into.