NO-ISSUE: Fix invalid StartLimitBurst value
Infinity is not a valid value in any version of systemd.

Setting it to 1000 means that the agent won't be prohibited from starting
as long as it doesn't restart more than 1000 times within 5 seconds, which
would be very difficult to achieve.

Greenboot is still responsible for ensuring the agent starts but systemd must
actually start it.