EDM-1659: Attempt to parallelize e2e tests
Each test gets its own device VM but shares the same flightctl backend
services.

This PR mostly consists of removing behaviors that indiscriminately
delete resources that could be used in other concurrent tests, since
they all share the same database.

There are 4 enrollmentrequest tests that are currently skipped because
the technique used to make the cli resources unique across tests does
not work with enrollmentrequests due to the certificate request
signature validation being tied to the ER name.

For the VM image, we can use the same technique used in the rollout
tests and create a qcow overlay image based on the original so that each
VM gets its own disk. This should minimize disk usage compared to
copying the disk for each parallel VM.