EDM-830: RBAC Testing an admin role
First test for RBAC.
Creating a role with creating,listing and getting flightctl resources and a a role binding to
a user with no prior permissions.

Making sure the user can:
Create, update, list, get a specific and delete a:
 - Device
 - Fleet
 - Repository

Create and list:
- Enrollment requests
- Events

Creating a role and a rolebinding in the default namespace and making sure the user can't perform any operations.
Creating a cluster role and a role binding and making sure the user can do the operations above.

Also add:
AddLabelsToYAML to add labels to a YAML (Should be modified in a followup patch)
Add a repostiory example file to test directory
Add GetRepository

Signed-off-by: Itzik Brown <itzikb@redhat.com>