EDM-2102: Render device spec upon label changes
The behavior of device rendering was broken because it skipped rendering
if the template version was already rendered. However, if a user updates
device labels which cause parameters to change, then the template
version stays the same but the device needs to be rendered again.

This fix removes that check. However, to keep the task idempotent, we
store a hash of the last rendered device spec in an annotation. When
starting the task, we compare the current hash to the one in the
annotation. This way, we're not bumping the renderedVersion if there
were no spec changes since the last render.