EDM-2098: SELinux - Allow read/write of bin_t domain
There are certain directories in /etc (e.g. /etc/profile.d) that uses
the file context of `bin_t` and the agent should be able to write there.

In general there is no reason the agent shouldn't be able to manage all
bin files on the system, even though generally this would be handled by
ostree/bootc.