EDM-2195: Use k8s job to create secrets. Do not render cleanup/db migration jobs for ACM.
Due to ACM's approach of running helm template + applying manifests manually, we can not use Secret resources directly.
We need to generate the Secret resources in runtime.
Due to the same reason, hooks also do not work so we skip cleanup and db migration jobs rendering when ACM is target platform.