EDM-2634: Add lychee-action to check doc links
This change adds a github workflow to check the links in our
documentation. It runs when creating PRs, periodically, and we should be
able to trigger a run manually.

I've run the checker, and it identified a number of broken links.  I've
taken a pass at fixing those broken links.