EDM-1332: Suppress bootc SELinux error
See https://github.com/bootc-dev/bootc/issues/1434 for the known issue on
bootc's side. We don't want to allow it because it might alter the behavior of bootc
in an unpredictable way, but we can suppress it in the avc logs.

(cherry picked from commit 691a0d0ad09d211cd0743c5bf0826255b538709c)