EDM-2748: ResourceSync should not overwrite annotations
Currently ResourceSync overwrites annotations with whatever is in the
fleet yaml in git because it's using an internal ctx. The ResourceSync
task should not be using that ctx because it's performing an action on
behalf of the user, and so should only be allowed to do what the user is
allowed to.

(cherry picked from commit a7426c918efe96e12b90afccb4219cc0003f465f)